pyproject.toml
[project]
name = "Samurai-Backend"
version = "0.16.0"
description = "Backend part of the Samurai project"
authors = [
    {name = "Pavlo Pohorieltsev", email = "hdydpavel@gmail.com"},
]
dependencies = [
    "fastapi[all]>=0.111.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.28",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pytz>=2024.1",
    "sqlmodel>=0.0.18",
    "alembic[tz]>=1.13.1",
    "asyncpg>=0.29.0",
    "websockets>=12.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "psycopg>=3.1.18",
    "opentelemetry-instrumentation-psycopg>=0.45b0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-logging>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
]
# opentelemetry-instrumentation-logging May be used, but not required for now.

requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "ruff>=0.4.5",
    "pre-commit>=3.7.0",
    "pyright>=1.1.357",
    "pylint[spelling]>=3.2.2",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
_echo_done={shell = "echo Done!"}
_perform_lint="ruff check . --fix"
lint={composite = ["_perform_lint", "_echo_done"], help="Executes ruff check, and echoes 'Done!'."}
format="ruff format ."
run_ruff={composite = ["format", "lint"], help="Runs formatter and linter on the code"}
run_server="uvicorn samurai_backend.app:app --reload"
create_admin="python src/samurai_backend/scripts/create_admin.py"
change_password="python src/samurai_backend/scripts/change_password.py"
create_permissions="python src/samurai_backend/scripts/create_permissions.py"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
extend-select = [
  "UP",
  "E",
  "W",
  "I",
  "C90",
  "N",
  "ANN",
  "ASYNC",
  "BLE",
  "FBT",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "FA",
  "ISC",
  "ICN",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "Q",
  "RET",
  "SLF",
  "SIM",
  "TID",
  "TCH",
  "ARG",
  "TD",
  "FIX",
  "ERA",
  "PL",
  "RUF",
]
ignore = [
  "ISC001",
  "N999",
  "ERA001",
  "RUF012",
  "ANN101",
  "ANN102",
  "FBT001",
  "FBT002",
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true


[tool.pyright]
include = [
  "src/samurai_backend",
]
exclude = [
  "src/samurai_backend/migrations",
]


[tool.pylint.MASTER]
ignore-paths = [
  "src/samurai_backend/migrations/"
]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 100
disable = [
  "missing-module-docstring",
  "missing-function-docstring",
  "missing-class-docstring",
  "logging-fstring-interpolation",
  "import-outside-toplevel",
]
